Key mapping
Changelog:
- Completed mapping to keys
- Made it so octave shifting is possible
- Started working on fix for sine popping effect (needs work still)

+++ MAGI/Assets/Scripts/GeneralSynth/FrequencyTable.cs
@@ -21,7 +21,7 @@ namespace GeneralSynth
         /// <summary>
         ///  This is the index of A4 in the standard frequency table
         /// </summary>
-        public float BaseKeyNumber => baseKeyNumber;
+        public int BaseKeyNumber => baseKeyNumber;
 
         /// <summary>
         ///  Calculate the frequency of a given key number
